int myRoot(double x)をdouble型に変更

+++ main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
 {
   double y;
   if( x <= 0 ){
